[add, update] 詳細画面の追加，計算クラスの作成
